chore(mobile): refactor error pages to not show error message and use the same component
