+++ apps/mobile/src/components/error/error.tsx
@@ -1,7 +1,8 @@
+import { useRefreshHandler } from '@/hooks/use-refresh-handler';
 import { t } from '@lingui/macro';
 import { Image } from 'expo-image';
 
-import { Box, Button, HasChildren, Text } from '@leather.io/ui/native';
+import { ArrowRotateClockwiseIcon, Box, Button, HasChildren, Text } from '@leather.io/ui/native';
 
 import { EmptyLayout } from '../loading/empty-layout';
 import { PageLayout } from '../page/page.layout';
@@ -14,12 +15,8 @@ function WidgetWrap({ children }: HasChildren) {
   );
 }
 
-interface ErrorProps {
-  error?: Error;
-  onRetry?: () => void;
-}
-
-export function Error({ error, onRetry }: ErrorProps) {
+export function Error() {
+  const { onRefresh } = useRefreshHandler();
   return (
     <PageLayout>
       <WidgetWrap>
@@ -37,31 +34,13 @@ export function Error({ error, onRetry }: ErrorProps) {
           <Text variant="heading03" textAlign="center" fontSize={32}>
             {t({ id: 'error_boundary.title', message: 'Something went wrong' })}
           </Text>
-          {onRetry ? (
-            <Button
-              onPress={onRetry}
-              title={t({ id: 'error_boundary.action', message: 'Try again' })}
-            />
-          ) : (
-            <>
-              <Text variant="label01" textAlign="center">
-                {t({ id: 'error_boundary.subtitle', message: 'Drag to refresh' })}
-              </Text>
 
-              {error && (
-                <Box
-                  width={218}
-                  height="auto"
-                  alignItems="flex-start"
-                  bg="ink.background-secondary"
-                >
-                  <Text variant="code" textAlign="center">
-                    {error.message}
-                  </Text>
-                </Box>
-              )}
-            </>
-          )}
+          <Button
+            onPress={() => onRefresh()}
+            icon={<ArrowRotateClockwiseIcon color="ink.background-primary" />}
+            buttonState="default"
+            title={t({ id: 'error_boundary.action', message: 'Refresh' })}
+          />
         </EmptyLayout>
       </WidgetWrap>
     </PageLayout>
