+++ apps/mobile/src/components/error/fetch-error.tsx
@@ -1,24 +1,6 @@
 import { t } from '@lingui/macro';
-import { Image } from 'expo-image';
 
 import { Box, Callout, CloudOffIcon, Text } from '@leather.io/ui/native';
-
-import { EmptyLayout } from '../loading/empty-layout';
-
-interface ErrorProps {
-  errorMessage?: string;
-}
-export function Error({ errorMessage }: ErrorProps) {
-  return (
-    <EmptyLayout image={<Image source={require('@/assets/stickers/net.png')} />}>
-      <Text variant="heading03">{t({ id: 'error.title', message: 'Something went wrong' })}</Text>
-      <Text variant="label01">
-        {t({ id: 'fetch-state-error.balance.subtitle', message: 'Pull this page to refresh' })}
-      </Text>
-      {errorMessage && <Text variant="code">{errorMessage}</Text>}
-    </EmptyLayout>
-  );
-}
 
 export function FetchErrorCallout() {
   return (
